+++ src/main/java/com/example/https/Entity/Department.java
@@ -1,5 +1,6 @@
 package com.example.https.Entity;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
