+++ src/main/java/com/example/https/HttpsApplication.java
@@ -2,8 +2,9 @@ package com.example.https;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = {SecurityAutoConfiguration.class})
 public class HttpsApplication {
 
     public static void main(String[] args) {
